requirements.txt
google-generativeai>=0.3.0
pytest>=7.0.0
gitpython
jira
python-dotenv>=1.0.0
numpy>=1.24.0
torch>=2.0.0
pytest-timeout
requests>=2.31.0
openpyxl>=3.1.2
pyautogen>=0.2.0
autogen>=0.9.6
openai>=1.0.0

llama-index
beautifulsoup4>=4.12.0

# Additional document processing libraries for multi-format input support
python-pptx>=0.6.21        # PowerPoint files
PyPDF2>=3.0.1              # PDF files (alternative: pymupdf)
xlrd>=2.0.1                # Legacy Excel files (.xls)